flaskapp/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Migrate==4.0.5
numpy==1.24.3
opencv-python==4.8.0
dlib==19.24.0